// src/app/modules/home/models/search.interface.ts
export interface ResponseI {
  incomplete_results: boolean;
  items: UsersI;
  total_count: number;
}
export interface UsersI {
  avatar_url: string;
  events_url: string;
  followers_url: string;
  following_url: string;
  gists_url: string;
  gravatar_id: string;
  html_url: string;
  id: number;
  login: string;
  node_id: string;
  organizations_url: string;
  received_events_url: string;
  repos_url: string;
  site_admin: boolean;
  starred_url: string;
  subscriptions_url: string;
  type: string;
  url: string;
}